More CameraTargetDistance math.
